+++ HISTORY.rst
@@ -90,3 +90,7 @@ History
 * Requirements upgraded.
 * Drop the use of some libraries: jsonpickle, twine.
 * Tox is installed in the Makefile when used.
+
+4.3.0 (2021-04-XX)
+==================
+* New parameter `retry_when_empty_result` for the `read`, `browse`, `get` and `request` methods to retry the query when resources are missing in the response.